requirements.txt
# File: requirements.txt
groq
python-dotenv
psutil
GitPython
